Implement parsing multiple amounts in single string

+++ word-indexer-bin/src/main.rs
@@ -44,14 +44,15 @@ fn main() {
 #[cfg(test)]
 pub mod integration_tests {
     use rstest::rstest;
+    use rust_decimal::Decimal;
     use word_indexer_lib::{models::Word, processor};
 
     #[rstest]
-    #[case("This has 1,003.28 as the single number", vec![1_003.28])]
-    #[case("This has 1,003.28; as the single number with trailing punctuation", vec![1_003.28])]
-    #[case("This has 1,003.28 and 837,291.37 as multiple numbers", vec![1_003.28, 837_291.37])]
-    #[case("This has some1,003.28within text", vec![1_003.28])]
-    fn monetary_amount(#[case] sentence: &str, #[case] expected: Vec<f64>) {
+    #[case("This has 1,003.28 as the single number", vec![Decimal::from_str_exact("1003.28").unwrap()])]
+    #[case("This has 1,003.28; as the single number with trailing punctuation", vec![Decimal::from_str_exact("1003.28").unwrap()])]
+    #[case("This has 1,003.28 and 837,291.37 as multiple numbers", vec![Decimal::from_str_exact("1003.28").unwrap(), Decimal::from_str_exact("837291.37").unwrap()])]
+    #[case("This has some1,003.28within text", vec![Decimal::from_str_exact("1003.28").unwrap()])]
+    fn monetary_amount(#[case] sentence: &str, #[case] expected: Vec<Decimal>) {
         let words: Vec<Word> = crate::create_words_from_sentence(sentence);
         let actual = processor::index_amounts(words);
         assert_eq!(expected, actual);
